refactor(web): extract EditPanel for LC
