+++ web/.vscode/settings.json
@@ -1,3 +1,10 @@
 {
-  "cSpell.words": ["Athanasiou", "defensemen", "FHFH", "Immer", "Postgrest"]
+  "cSpell.words": [
+    "Athanasiou",
+    "clickaway",
+    "defensemen",
+    "FHFH",
+    "Immer",
+    "Postgrest"
+  ]
 }
