+++ web/pages/lines/edit/[id].module.scss
@@ -1,95 +1,6 @@
 @import "styles/vars.scss";
 
 .container {
-  .top {
-    display: flex;
-    justify-content: space-between;
-    align-items: flex-end;
-    .user {
-      p {
-        margin: 0;
-      }
-      a {
-        &:hover {
-          color: $primary-color;
-          cursor: pointer;
-        }
-      }
-    }
-    .time {
-      font-style: italic;
-      font-weight: 800;
-      font-size: 1.125rem;
-      span {
-        color: rgba(255, 255, 255, 0.5);
-      }
-    }
-  }
-  .mainContent {
-    display: flex;
-    flex-direction: column;
-
-    .line {
-      margin-bottom: 12px;
-    }
-
-    .sourceUrl {
-      display: flex;
-      align-items: center;
-      gap: 12px;
-      margin-top: 20px;
-
-      label {
-        color: rgba(255, 255, 255, 0.5);
-        font-style: italic;
-        font-weight: 800;
-        font-size: 1.125rem;
-      }
-      input {
-        flex: 1;
-        background: rgba(255, 255, 255, 0.1);
-        border: 1px solid rgba(255, 255, 255, 0.25);
-        border-radius: 4px;
-        height: 29px;
-        color: white;
-        padding-left: 0.5em;
-      }
-    }
-
-    .action {
-      margin-top: 20px;
-      display: flex;
-      justify-content: center;
-      align-items: center;
-      gap: 10px;
-      button {
-        all: unset;
-        background: rgba(7, 170, 226, 0.85);
-        border: 1px solid rgba(255, 255, 255, 0.85);
-        padding: 0.3em 2em;
-        border-radius: 4px;
-        font-style: italic;
-        font-weight: 800;
-        font-size: 18px;
-        letter-spacing: 0.1em;
-        text-transform: uppercase;
-        color: rgba(255, 255, 255, 0.9);
-
-        display: flex;
-        justify-content: center;
-        align-items: center;
-        &:hover {
-          cursor: pointer;
-          opacity: 0.9;
-        }
-      }
-
-      .delete {
-        background-color: red;
-        color: white;
-      }
-    }
-  }
 }
 
 @media only screen and (min-width: $screen-large) {
@@ -97,44 +8,5 @@
     min-width: 1024px;
     margin: 0 auto;
     padding-bottom: 8rem;
-
-    .mainContent {
-      border: 1px solid white;
-      padding: 26px 37px 28px 28px;
-      background: rgba(255, 255, 255, 0.05);
-      flex-direction: row;
-      gap: 1rem;
-      .categoryTitle {
-        padding-left: 54px;
-      }
-      .line:not(:last-of-type) {
-        margin-bottom: 64px;
-      }
-
-      .sourceUrl {
-        padding-left: 54px;
-        margin-top: 40px;
-
-        label {
-          color: rgba(255, 255, 255, 0.5);
-          font-style: italic;
-          font-weight: 800;
-          font-size: 1.125rem;
-        }
-        input {
-          flex: 1;
-          background: rgba(255, 255, 255, 0.1);
-          border: 1px solid rgba(255, 255, 255, 0.25);
-          border-radius: 4px;
-          height: 29px;
-          color: white;
-        }
-      }
-
-      .action {
-        padding-left: 54px;
-        margin-top: 37px;
-      }
-    }
   }
 }
